src/main/java/ru/otus/homework/service/PostServiceImpl.java
package ru.otus.homework.service;

import lombok.SneakyThrows;
import org.springframework.stereotype.Service;
import ru.otus.homework.model.Book;
import ru.otus.homework.model.Post;
import ru.otus.homework.repository.BookInfoRepositoryJpa;
import ru.otus.homework.repository.PostRepositoryJpa;

import java.io.UnsupportedEncodingException;
import java.util.List;
import java.util.Optional;

@Service
public class PostServiceImpl implements PostService {

    private final CommunicationService communicationService;
    private final BookInfoRepositoryJpa bookInfoRepositoryJpa;
    private final PostRepositoryJpa postRepositoryJpa;

    public PostServiceImpl(
            CommunicationService communicationService,
            BookInfoRepositoryJpa bookInfoRepositoryJpa,
            PostRepositoryJpa postRepositoryJpa
    ) {
        this.communicationService = communicationService;
        this.bookInfoRepositoryJpa = bookInfoRepositoryJpa;
        this.postRepositoryJpa = postRepositoryJpa;
    }

    @SneakyThrows
    @Override
    public void insertPostByBook() {
        Book book = getBook();
        if (book != null) {
            String description = communicationService.getUserInputString(
                    "Введите комментарий к книге",
                    "Некорректный комментарий! Введите еще раз",
                    "[^.]+"
            );
            Post post = postRepositoryJpa.save(new Post(0, book, description));
            if (post != null) {
                communicationService.showMessage("inserted: " + post.toString());
            } else {
                communicationService.showMessage("the book was not insert");
            }
        } else {
            communicationService.showMessage("the book was not found");
        }
    }

    @SneakyThrows
    @Override
    public void deletePostsByBook() {
        Book book = getBook();
        if (book != null) {
            int countPost = postRepositoryJpa.deleteByBook(book);
            communicationService.showMessage(String.format("deleted %d posts", countPost));
        } else {
            communicationService.showMessage("the book was not found");
        }
    }

    @SneakyThrows
    @Override
    public void getPostsByBook() {
        Book book = getBook();
        if (book != null) {
            List<Post> posts = postRepositoryJpa.findByBook(book);
            if (posts.size() > 0) {
                for (Post post : posts) {
                    communicationService.showMessage(post.toString());
                }
            } else {
                communicationService.showMessage("the posts was not found");
            }
        } else {
            communicationService.showMessage("the book was not found");
        }
    }

    private Book getBook() throws UnsupportedEncodingException {
        long bookId = Long.parseLong(
                communicationService.getUserInputString(
                        "Введите идентификатор книги",
                        "Некорректный идентификатор! Введите еще раз",
                        "[^d]+"
                )
        );
        Optional<Book> optionalBook = bookInfoRepositoryJpa.findById(bookId);
        return optionalBook.orElse(null);
    }
}
